Made a "dumb" reference to original doc
Bringing in a foreign key, handle updates in elastic and new format was
too big of a task. Simple changed the way we delete/update/insert new
docs. The new way is to set the "source" as allowed_clients.name by
the client-reference in jobs. This should not break previous behavior.